# openapi.yaml
openapi: "3.1.0"
info:
  title: LeanStock API
  version: "2.0.0"
  description: |
    Multi-tenant inventory management system with atomic transfers,
    dead stock decay, predictive reordering, reservation system,
    and full supplier/purchase-order workflows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://leanstock.onrender.com
    description: Production (Render.com)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user and company
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "ceo@mycompany.com"
              username: "johndoe"
              password: "secret123"
              tenant_name: "MyCompany"
      responses:
        "201":
          description: Registered — verification email sent
        "409":
          description: Email or username already taken
        "422":
          description: Password too short
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "ceo@mycompany.com"
              password: "secret123"
      responses:
        "200":
          description: Access + refresh tokens
          content:
            application/json:
              example:
                access_token: "eyJ..."
                refresh_token: "abc..."
                token_type: "bearer"
        "401":
          description: Invalid credentials
        "403":
          description: Email not verified or account deactivated
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              refresh_token: "abc..."
      responses:
        "200":
          description: New access token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              refresh_token: "abc..."
      responses:
        "200":
          description: Logged out
  /auth/verify-email/{token}:
    get:
      tags: [Auth]
      summary: Verify email via token from link
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email verified
        "400":
          description: Token expired
        "404":
          description: Invalid token
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "ceo@mycompany.com"
      responses:
        "200":
          description: Always 200 (security — does not reveal if email exists)
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Complete password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              token: "reset-token-from-email"
              new_password: "newpassword123"
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Invalid or expired token
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users in tenant (admin only)
      responses:
        "200":
          description: User list
    post:
      tags: [Admin]
      summary: Create user and send invite email (admin only)
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "manager@company.com"
              username: "johnmanager"
              password: "secret123"
              role: "manager"
      responses:
        "201":
          description: User created, invite email sent
  /admin/users/{user_id}/role:
    put:
      tags: [Admin]
      summary: Update user role (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              role: "manager"
      responses:
        "200":
          description: Role updated
  /admin/users/{user_id}:
    delete:
      tags: [Admin]
      summary: Deactivate user (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
...
  /admin/trigger-decay:
    post:
      tags: [Admin]
      summary: Manually trigger dead stock decay Celery task
      parameters:
        - name: decay_percent
          in: query
          schema:
            type: number
            default: 10.0
      responses:
        "200":
          description: Task queued
...